Simplified kblib::poly_obj and fixed UB.
kblib::poly_obj now no longer uses std::launder or reinterpret_cast
internally.

I tried porting it to C++14 as well, but it makes significant use of
constexpr if, so it's still C++17-only and probably will remain that
way. (Also the type traits broke in C++14 for some reason).