Added magic_enumerate (pending rename), which is an enumerate iterator wrapper (it wraps an iterator to produce an index on dereference) with "magic value semantics".
Currently it is tested just as well as the regular (reference-only)
enumerate. I'll probably rename the old one to enumerate_ref and the new
one to enumerate. (The old one is simpler and almost certainly produces
better assembly, so it's probably a better choice when that matters, but
I haven't done any performance testing yet.)