Improved magic_enumerate to now properly decay away const on copy. This was both uglier and not as terrible as I expected it would be. Also, I moved the tests to iterators.cpp.
Unfortunately, the constraints I'm working with mean that I had to abuse
volatile (and const_cast) a lot. It is therefore unsupported to ever
actually declare a volatile enumeration<T>, but that's hardly unusual.

I have added quite a few more tests and it seems that everything checks
out, though, so I currently believe that my abuse of volatile is okay,
especially in a function with 'magic' in its name.