Added beginnings of a fairly complex sort functionality. Currently, only a basic insertion_sort and a slightly less basic out-of-place insertion_sort_copy are implemented. insertion_sort is currently broken, see test failure. insertion_sort_copy fully working, performance not yet tested. Lots of minor fixes, largely for typos in documentation and missing constexpr on functions.
adaptive_insertion_sort_copy is probably pretty unnecessary, but at the
cost of Θ(sqrt(n)) extra computation, it avoids quadratic time for
reverse-sorted inputs. Therefore, its best-case is "almost sorted input,
in either direction".

Θ(sqrt(n)) is smaller than O(n) and especially smaller than O(n^2), so
technically its asymptotic bounds are compatible with the standard
insertion_sort description. I chose to write the specification in big
theta notation purely because I wanted to make it clear that its best-
case scenario is not purely one-pass like a normal insertion sort.

It is easily possible to construct data which provokes worst-case
performance even with the extra pass, because of the algorithm I used,
but because its worst case is asymptotically equal to its average case,
I don't think that matters. The worst-case is no longer 'plausible',
which is what I wanted.

The factor of sqrt(n) is not rigorously chosen, but is based on optimal
choice strategy: the first sqrt(n) elements of a collection tell you
with reasonable confidence approximately what the overall shape of the
collection is.