Added pointer-to-member-data operator overload to poly_obj. pointer-to- member-function overload poses challenges, though. TODO: documentation for them, if I even keep them.
Moved copy_const to fakestd.h. I forgot where it was before, but
copy_const_t was already there. Perhaps I split up a patch accidentally.

In preparation for C++20, a helper type zero_t was added to simple.h
with an implicit consteval constructor that only accepts 0 as an
argument (Otherwise it's a guaranteed compile-time error) and which is
implicitly convertible to anything that a literal zero is convertible
to, as well as to nullptr_t. It's hidden behind an #if right now.

Apparently I just never formatted variant.h? Well it's formatted now.
Pardon the large diff.