Fixed several issues with FNV_hash that I hadn't noticed before, mainly pertaining to padding and empty types/bools.
Added a hashing optimization for sequences of trivially-hashable types
to hash them as a sequence of bytes instead of each element
individually.

Changed the definitions of is_linear_container to be more C++14-
compliant.

Fixed an ambiguous call error in consume_iterator.

Changed the definition of a 'trivially hashable' type. The old behavior
would cause inconsistent hashes for types with padding. Now, In C++14,
only integral types without padding can be hashed as primitives (this is
the best I can do in C++14). In C++17, any trivially copyable type with
no padding may be hashed as a primitive.

Rationale: Using a trivial hash for empty types, and using the trivial
hash for bools would lead to serious collisions in the case of hashing
sequences of such types (as the hashes of every sequence of empty types
would be identical no matter the length, and the bool hash was
effectively just the xor fold of the input).

TODO: test for C++14 compliance and verify that the hashing actually
works (currently, I only know that it compiles with everything I've
thrown at it).