Fixed a bunch of errors found by decltype(auto) and wreien on the C++ Help Discord server, including several GCC compatibility bugs.
Added basic tests for get_max_n and get_max_n_old.

Finally cleaned up the last of the nonsense information that was printed
instead of being automatically checked. The replacement in the case of
the poly_obj tests is, however, somewhat less than ideal. (It just
writes the same information to a global and then verifies it.)