Get kblib working on GCC/libstdc++
Also small improvements to the build config.
This commit unfortunately also disables sanitizers because of a library
issue.

This commit is the result of work over an extended period because I once
again forgot to commit. It may or may not work by itself, as I separated
out the larger features into the following commits.